--- src/main/java/com/rakovets/course/java/core/practice/concurrency/consumerWorkResult
@@ -1,84 +1,0 @@
-
-2021-05-19T22:11:33.317949400 Consumer - I slept 5 seconds
-2021-05-19T22:12:47.668953100 Consumer - I slept 5 seconds
-2021-05-19T22:12:51.687552400 Consumer - I slept 4 seconds
-2021-05-19T22:12:59.695624200 Consumer - I slept 8 seconds
-2021-05-19T22:16:41.071788700 Consumer - I slept 5 seconds
-2021-05-19T22:16:50.101125300 Consumer - I slept 9 seconds
-2021-05-19T22:16:53.106265900 Consumer - I slept 3 seconds
-2021-05-19T22:17:54.379372200 Consumer - I slept 5 seconds
-2021-05-19T22:18:02.404435400 Consumer - I slept 8 seconds
-2021-05-19T22:18:06.404656500 Consumer - I slept 4 seconds
-2021-05-19T22:33:23.969719400 Consumer ...
-2021-05-19T22:33:24.991735100 Consumer ...
-2021-05-19T22:33:26.005749600 Consumer ...
-2021-05-19T22:33:27.008025100 Consumer ...
-2021-05-19T22:33:28.013517800 Consumer ...
-2021-05-19T22:33:29.015851600 Consumer ...
-2021-05-19T22:33:30.026346600 Consumer ...
-2021-05-19T22:33:31.036307900 Consumer ...
-2021-05-19T22:33:32.045961800 Consumer ...
-2021-05-19T22:33:33.049102 Consumer ...
-2021-05-19T22:33:34.055680700 Consumer ...
-2021-05-19T22:33:42.065087600 Consumer - I slept 8 seconds
-2021-05-20T00:23:03.895462900 Consumer ...
-2021-05-20T00:23:04.902849300 Consumer ...
-2021-05-20T00:23:05.915709400 Consumer ...
-2021-05-20T00:23:06.928967100 Consumer ...
-2021-05-20T00:23:07.943199300 Consumer ...
-2021-05-20T00:23:08.956154700 Consumer ...
-2021-05-20T00:23:09.965434 Consumer ...
-2021-05-20T00:23:10.971808900 Consumer ...
-2021-05-20T00:23:11.983015700 Consumer ...
-2021-05-20T00:23:12.995345300 Consumer ...
-2021-05-20T00:23:20.999109300 Consumer - I slept 8 seconds
-2021-05-20T00:23:24.025659400 Consumer - I slept 3 seconds
-2021-05-20T00:23:32.434094300 Consumer ...
-2021-05-20T00:23:33.455119300 Consumer ...
-2021-05-20T00:23:34.464773300 Consumer ...
-2021-05-20T00:23:35.469082 Consumer ...
-2021-05-20T00:23:40.482690200 Consumer - I slept 5 seconds
-2021-05-20T00:23:48.510703900 Consumer - I slept 8 seconds
-2021-05-20T00:23:52.515457300 Consumer - I slept 4 seconds
-2021-05-20T00:23:55.517441600 Consumer - I slept 3 seconds
-2021-05-20T00:25:53.881090900 Consumer ...
-2021-05-20T00:25:54.896679900 Consumer ...
-2021-05-20T00:25:55.906496500 Consumer ...
-2021-05-20T00:25:56.917882800 Consumer ...
-2021-05-20T00:25:57.926988 Consumer ...
-2021-05-20T00:25:58.937068500 Consumer ...
-2021-05-20T00:26:03.942455500 Consumer - I slept 5 seconds
-2021-05-20T00:26:11.971646200 Consumer - I slept 8 seconds
-2021-05-20T00:26:15.985286 Consumer - I slept 4 seconds
-2021-05-20T00:26:18.998598900 Consumer - I slept 3 seconds
-2021-05-20T00:26:21.001951200 Consumer - I slept 2 seconds
-2021-05-20T00:28:36.537567900 Consumer ...
-2021-05-20T00:28:37.554409700 Consumer ...
-2021-05-20T00:28:38.566339600 Consumer ...
-2021-05-20T00:28:43.568892100 Consumer - I slept 5 seconds
-2021-05-20T00:28:47.592275900 Consumer - I slept 4 seconds
-2021-05-20T00:28:48.602879900 Consumer - I slept 1 seconds
-2021-05-20T00:31:43.366901400 Consumer2 ...
-2021-05-20T00:31:43.366901400 Consumer3 ...
-2021-05-20T00:31:43.366901400 Consumer ...
-2021-05-20T00:31:44.383711900 Consumer2 ...
-2021-05-20T00:31:44.383711900 Consumer3 ...
-2021-05-20T00:31:44.383711900 Consumer ...
-2021-05-20T00:31:45.391642300 Consumer ...
-2021-05-20T00:31:45.391642300 Consumer3 ...
-2021-05-20T00:31:45.391642300 Consumer2 ...
-2021-05-20T00:31:46.403041100 Consumer2 ...
-2021-05-20T00:31:46.403041100 Consumer ...
-2021-05-20T00:31:46.403041100 Consumer3 ...
-2021-05-20T00:31:47.413250900 Consumer2 ...
-2021-05-20T00:31:47.413250900 Consumer ...
-2021-05-20T00:31:47.413250900 Consumer3 ...
-2021-05-20T00:31:48.422714 Consumer ...
-2021-05-20T00:31:48.422714 Consumer2 ...
-2021-05-20T00:31:48.422714 Consumer3 ...
-2021-05-20T00:31:53.439925800 Consumer - I slept 5 seconds
-2021-05-20T00:31:53.439925800 Consumer3 - I slept 5 seconds
-2021-05-20T00:31:53.439925800 Consumer2 - I slept 5 seconds
-2021-05-20T00:31:59.473763 Consumer3 - I slept 6 seconds
-2021-05-20T00:31:59.473763 Consumer2 - I slept 6 seconds
-2021-05-20T00:31:59.473763 Consumer - I slept 6 seconds
